clean up C-part detectcoll
