+++ src/collisiondetection/detectcoll.h
